+++ src/screens/Tab/LedgerScreen.tsx
@@ -1,7 +1,7 @@
 import RecentTransactionOverview from 'components/screens/LedgerScreen/RecentTransactionOverview';
 import AvailableFundsOverview from 'components/screens/LedgerScreen/AvailableFundsOverview';
 import IncomeStatementOverview from 'components/screens/LedgerScreen/IncomeStatementOverview';
-import FinancialStatusOverview from 'components/screens/LedgerScreen/FInancialStatusOverview';
+import FinancialStatusOverview from 'components/screens/LedgerScreen/FinancialStatusOverview';
 import ScreenContainer from 'components/@common/ScreenContainer';
 import Divider from 'components/@common/Divider';
